requirements.txt
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.6.0
cryptography==2.9
ecdsa==0.15
Flask==1.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jose==1.0.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-spec-validator==0.2.8
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.3.6
PyJWT==1.7.1
pyrsistent==0.15.7
python-jose==3.1.0
python-jose-cryptodome==1.3.2
PyYAML==5.3
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
swagger-ui-bundle==0.0.6
urllib3==1.25.8
Werkzeug==1.0.0
zipp==3.1.0
